Change markup and styling for Check your answers page
Switch markup from using a table to using a description list
and improve styling to be clearer and responsive.

Visual changes:
* Don't use the full page width if the items are short
* Remove the bold from the change link
* Make the key bold
* Make key and value wrap on smaller screens
* The width of the first column can be controlled via
  modifiers `cya-questions-short` or `cya-questions-long`,
  if there is no modifier, 'columns' automatically adjust
  their widths to their content

The description list uses a div around each dt/dd grouping.
That is currently invalid but
a) in the HTML 5.2 Working Draft [1]
b) supported in all the browsers (even IE6)

[1] https://www.w3.org/TR/html52/grouping-content.html#the-dl-element